Added OpenApi3 and Swagger, added openapi.yaml doc

+++ src/main/java/ua/lozychenko/nonogram/data/entity/util/GameStats.java
@@ -5,5 +5,5 @@ public interface GameStats {
 
     String getEmail();
 
-    Double getScore();
+    Integer getScore();
 }
